+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog for languages
 
-## 26 October 2016 - Initial release
+## 2016-11-06
+  * Feedback home page
+  * Fix en pluralization
+
+## 2016-11-02
+  * Word count challenge
+
+## 2016-10-26 - Initial release